fix: resolve different things
